feat: complete backend with tests for all controllers, services and repositories
- Implement comprehensive tests for Admin, City, MobilityService, Offer, User controllers
- Add service layer tests with mock repositories
- Include repository tests with H2 database
- Configure test security exclusions
- Fix Java version compatibility issues

+++ backend/src/main/java/com/mobility/mobility_backend/controller/CityController.java
@@ -1,23 +1,30 @@
 package com.mobility.mobility_backend.controller;
- 
-import com.mobility.mobility_backend.dto.CityDTO;
-import com.mobility.mobility_backend.service.CityService;
+
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import java.util.List;
-import java.util.Optional;
+
+import com.mobility.mobility_backend.dto.CityDTO;
+import com.mobility.mobility_backend.service.CityService;
 
 @RestController
 @RequestMapping("/api/cities")  // ✅ Tous les endpoints commencent par /api/cities
 public class CityController {
-	
+
 	 @Autowired
 	    private CityService cityService;
-	 
-	 
+
+
 	 /**
 	     * Crée une nouvelle ville
 	     * POST /api/cities
@@ -31,8 +38,8 @@ public class CityController {
 	            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
 	        }
 	    }
-	 
-	 
+
+
 	   /**
 	     * Récupère toutes les villes
 	     * GET /api/cities
@@ -42,19 +49,19 @@ public class CityController {
 	        List<CityDTO> cities = cityService.getAllCities();
 	        return new ResponseEntity<>(cities, HttpStatus.OK);
 	    }
-	    
-	    
+
+
 	    /**
 	     * Récupère une ville par son ID
 	     * GET /api/cities/{id}
 	     */
 	    @GetMapping("/{id}")
-	    public ResponseEntity<CityDTO> getCityById(@PathVariable Long id) {
+	    public ResponseEntity<CityDTO> getCityById(@PathVariable Integer id) {
 	        Optional<CityDTO> city = cityService.getCityById(id);
 	        return city.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
 	                  .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
 	    }
-	    
+
 	    /**
 	     * Récupère une ville par son nom
 	     * GET /api/cities/name/{name}
@@ -65,17 +72,17 @@ public class CityController {
 	        return city.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
 	                  .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
 	    }
-	    
+
 	    /**
 	     * Supprime une ville par son ID
 	     * DELETE /api/cities/{id}
 	     */
 	    @DeleteMapping("/{id}")
-	    public ResponseEntity<Void> deleteCity(@PathVariable Long id) {
+	    public ResponseEntity<Void> deleteCity(@PathVariable Integer id) {
 	        boolean deleted = cityService.deleteCity(id);
-	        return deleted ? new ResponseEntity<>(HttpStatus.NO_CONTENT) 
+	        return deleted ? new ResponseEntity<>(HttpStatus.NO_CONTENT)
 	                      : new ResponseEntity<>(HttpStatus.NOT_FOUND);
 	    }
-	    
+
 
 }
