+++ backend/src/main/java/com/mobility/mobility_backend/repository/MobilityServiceRepository.java
@@ -1,15 +1,16 @@
 package com.mobility.mobility_backend.repository;
 
-import com.mobility.mobility_backend.entity.MobilityService;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+import com.mobility.mobility_backend.entity.MobilityService;
 
 @Repository
-public interface MobilityServiceRepository extends JpaRepository<MobilityService, Long> {
-    
+public interface MobilityServiceRepository extends JpaRepository<MobilityService, Integer> {
+
     Optional<MobilityService> findByName(String name);
-    
+
     boolean existsByName(String name);
 }
